@@ -1,4 +1,4 @@
-class CreateQuestionAnwers < ActiveRecord::Migration
+class CreateQuestionAnwers < ActiveRecord::Migration[4.2]
   def change
     create_table :question_answers do |t|
 
