@@ -1,4 +1,4 @@
-class AddingValueToQuestionAnser < ActiveRecord::Migration
+class AddingValueToQuestionAnser < ActiveRecord::Migration[4.2]
   def change
     add_column :question_answers, :value, :text
   end
